requirements.txt
requests>=2.28.1
streamlit>=1.27.0
pandas>=2.0.0
numpy>=1.24.0
vaderSentiment>=3.3.2
scikit-learn>=1.3.0
xgboost>=1.7.0
joblib>=1.3.0
python-dotenv>=1.0.0